pyproject.toml
[project]
name = "aanvraagapp"
version = "0.1.0"
description = "Write proposals for whatever blazingly fast"
authors = [
    {name = "mark@markdewijk.nl"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "sqlalchemy (>=2.0.42,<3.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "redis[hiredis] (>=6.3.0,<7.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "trafilatura (>=2.0.0,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pgvector (>=0.4.1,<0.5.0)",
    "google-genai (>=1.32.0,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "beautifulsoup4 (>=4.13.5,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
debugpy = "^1.8.15"
uvicorn = "^0.35.0"
ruff = "^0.12.8"
pytest = "^8.4.2"
pytest-asyncio = "0.21.2"

[tool.pytest.ini_options]
asyncio_mode = "auto"